+++ src/main/java/edugate/demo/repositories/CourseRealizationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import edugate.demo.model.CourseRealization;
 
-public interface CourseRealizationRepository extends JpaRepository<CourseRealization,Long>
+public interface CourseRealizationRepository extends JpaRepository<CourseRealization,Integer>
 {
 }
