+++ src/main/java/edugate/demo/repositories/FileRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import edugate.demo.model.File;
 
-public interface FileRepository extends JpaRepository<File,Long> {
+public interface FileRepository extends JpaRepository<File,Integer> {
 }
